Fixes a bug where focus isn't restored if the focused content is not a deep descendant of the overlay. (...)
Before, if the user was using a <slot> to distribute content into an element
having IronOverlayBehavior instead of directly placing the content as a
shadow-including descendant of the overlay...

```html
<outer-component>
  #shadow-root
    <element-with-iron-overlay-behavior>
      <slot name="overlay-content"></slot>
    </element-with-iron-overlay-behavior>
  #/shadow-root
  <div slot="overlay-content">
    This div is distributed into the overlay but isn't a shadow-including
    descendant of it. Here's a focusable element: <input>
  </div>
</outer-component>
```

...then, when the overlay was closing, it would decide not to restore focus
because the criteria for restoring focus was "no element is focused or the
focused element is a shadow-including descendant of this element".

This PR changes the "a shadow-including descendant of this element" part of
that criteria to "a descendant of this element in the composed (i.e. flattened)
tree".